requirements.txt
jupyterhub
oauthenticator
ruamel.yaml
jsonschema
jinja2